requirements.txt
apache-airflow-providers-postgres
structlog==23.2.0
psycopg2-binary
pandas
matplotlib
scikit-learn
joblib